// my.replika.ai/static/js/types/enums.ts
export enum MetricsEvents {
  Launched = 'Application launched',
  Upvote = 'Upvote',
  Downvote = 'Downvote',
  AccountDeleted = 'Account deleted',
  NeedHelpButtonClicked = 'Need help - Button clicked',
  NeedHelpOptionSelected = 'Need help - Option selected',
  NeedHelpHotlineCalled = 'Need help - Hotline called',
  NeedHelpLifelineChatClicked = 'Need help - Lifeline chat clicked',
  NeedHelpWebsiteClicked = 'Need help - Website clicked',
  ExperimentStarted = 'Experiment Started',
  ExperimentEnded = 'Experiment Ended',
  ForgotPassword = 'Forgot password',
  MoodAdded = 'Mood added',
  MoodEdited = 'Mood edited',
  MoodDeleted = 'Mood deleted',
  MoodSliderCalled = 'Mood slider called',
  MoodOpened = 'Mood opened',
  MoodSliderDisplayed = 'Mood slider displayed',
  MoodSliderSkipped = 'Mood slider skipped',
  WelcomeScreenOpened = 'Welcome screen opened',
  SignupScreenOpened = 'Signup screen opened',
  LoginScreenOpened = 'Login screen opened',
  TrackOpened = 'Track opened',
  CategoryOpened = 'Category opened',
  SpotlightOpened = 'Spotlight opened',
  ReplikaImageSet = 'Replika image set',
  CreditsOpened = 'Credits opened',
  HelpCenterOpened = 'Help center opened',
  SettingsOpened = 'Settings opened',
  PrivacyPolicyOpened = 'Privacy policy opened',
  TermsOfServiceOpend = 'Terms of service opened',
  PhraseShown = 'Phrase shown',
  UserAccountSet = 'User account set',
  ReplikaProfileSet = 'Replika profile set',
  ReplikaNameAdded = 'Replika name added',
  TestimonialShown = 'Testimonial shown',
  FirstOnboardingPhraseShown = '1st onboarding phrase shown',
  MeetButtonShown = 'Meet button shown',
  MeetButtonPressed = 'Meet button pressed',
  RelationshipStatusOpened = 'Relationship status opened',
  RelationshipStatusClosed = 'Relationship status closed',
  RelationshipStatusChanged = 'Relationship status changed',
  RelationshipStatusAttemptToChange = 'Relationship status attempt to change',
  SubscriptionWebPromptShown = 'Subscription web prompt shown',
  SubscriptionWebPromptClicked = 'Subscription web prompt clicked',
  SubscriptionExtendedTrialShown = 'Subscription extended trial shown',
  PageIsAutoTranslated = 'Page is auto-translated',
  SavedChatItemsPopupOpened = 'Save chat elements: pop-up opened',
  SavedChatItemsSaveButtonPressed = 'Save chat elements: save button pressed',
  SavedChatItemsCancelButtonPressed = 'Save chat elements: cancel button pressed',
  SavedChatItemsPopupEditNamePressed = 'Save chat elements: Pop-up edit name pressed',
  SavedChatItemsLibraryEditNamePressed = 'Save chat elements: Library edit name pressed',
  SavedChatItemOpened = 'Save chat elements: Element opened',
  SavedChatItemsEmptyStateButtonPressed = 'Save chat elements: Empty state pressed',
}

export enum Themes {
  Light = 'LIGHT',
  Dark = 'DARK',
}

export enum Genders {
  Female = 'female',
  NonBinary = 'non-binary',
  Male = 'male',
}

export enum Pronouns {
  She = 'she',
  He = 'he',
  They = 'they',
}

export enum Routes {
  Signup = '/signup',
  SignupCreateAccount = '/signup/create-account',
  SignupChooseYourAIFriend = '/signup/choose-your-ai-friend',
  SignupCustomizeYourReplika = '/signup/customize-your-replika',
  SignupChooseReplikaName = '/signup/choose-replika-name',
  SignupAlmostThere = '/signup/almost-there',
  Login = '/login',
  Main = '/',
  ConfirmSignup = '/confirm-signup',
  DeleteAccount = '/delete-account',
  RecoverPassword = '/recover-password',
  VerifyEmail = '/verify-email',
  Conversations = '/conversations',
  Conversation = '/conversations/:category/:trackId',
  ConversationCategory = '/conversations/:category',
  Intro = '/intro',
  Stats = '/stats',
  ChooseAvatar = '/choose-avatar',
  SavedChatItems = '/saved',
  SavedChatItemsTab = '/saved/:tab',
  SavedChatItem = '/saved/:tab/:itemId',
  Memory = '/memory',
  MemoryRecord = '/memory/:memoryType/:memoryId',
}

export enum Dialogs {
  Feedback = 'Feedback',
  EditProfile = 'EditProfile',
  EditReplika = 'EditReplika',
  AccountSettings = 'AccountSettings',
  EditPassword = 'EditPassword',
  DropMission = 'DropMission',
  Badge = 'Badge',
  EditMood = 'EditMood',
  IntroMeetReplika = 'IntroMeetReplika',
  GetHelp = 'GetHelp',
  SubscriptionReboarding = 'SubscriptionReboarding',
  Subscription = 'Subscription',
  EditRelationshipStatus = 'EditRelationshipStatus',
  ReboardingAddPronoun = 'ReboardingAddPronoun',
  RewardPopup = 'RewardPopup',
  CustomizeAvatar = 'CustomizeAvatar',
  Reaction = 'Reaction',
  EditSavedChatItem = 'EditSavedChatItem',
  NavigationPopup = 'NavigationPopup',
  CancelSubscription = 'CancelSubscription',
}

export enum IntroStep {
  Hi = 'Hi',
  YouCanTalkToMe = 'YouCanTalkToMe',
  Conversations = 'Conversations',
  MeetReplika = 'MeetReplika',
  FinishIntro = 'FinishIntro',
}